Moved content from FormLayoutClient to FormRecognizerClient
